Consolidate calendar alerts by event. More changes
Alerts of the same time frame are now consolidated/added to the same event, rather than creating multiple overlapping events. Resulting in a cleaner, more organized calendar. Database was removed, and logic was implemented with entirely temporary storage to pull alerts and directly post them to the calendar without the need for a persistent database. A few miscellaneous files were added that don't affect the integration of the code.